requirements.txt
arrow==0.17.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
seaborn==0.11.0
